refactor: global 패키지 구성
