+++ backend/src/main/java/com/woowacourse/momo/auth/config/AuthInterceptor.java
@@ -14,8 +14,8 @@ import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.auth.support.AuthorizationExtractor;
 import com.woowacourse.momo.auth.support.JwtTokenProvider;
-import com.woowacourse.momo.globalException.exception.ErrorCode;
-import com.woowacourse.momo.globalException.exception.MomoException;
+import com.woowacourse.momo.global.exception.exception.ErrorCode;
+import com.woowacourse.momo.global.exception.exception.MomoException;
 
 @RequiredArgsConstructor
 public class AuthInterceptor implements HandlerInterceptor {
