+++ backend/src/main/java/com/woowacourse/momo/auth/support/google/GoogleConnector.java
@@ -1,7 +1,7 @@
 package com.woowacourse.momo.auth.support.google;
 
-import static com.woowacourse.momo.globalException.exception.ErrorCode.OAUTH_ACCESS_TOKEN_REQUEST_FAILED_BY_NON_2XX_STATUS;
-import static com.woowacourse.momo.globalException.exception.ErrorCode.OAUTH_ACCESS_TOKEN_REQUEST_FAILED_BY_NON_EXIST_BODY;
+import static com.woowacourse.momo.global.exception.exception.ErrorCode.OAUTH_ACCESS_TOKEN_REQUEST_FAILED_BY_NON_2XX_STATUS;
+import static com.woowacourse.momo.global.exception.exception.ErrorCode.OAUTH_ACCESS_TOKEN_REQUEST_FAILED_BY_NON_EXIST_BODY;
 
 import java.util.Collections;
 import java.util.Optional;
@@ -22,7 +22,7 @@ import lombok.RequiredArgsConstructor;
 
 import com.woowacourse.momo.auth.support.google.dto.GoogleTokenResponse;
 import com.woowacourse.momo.auth.support.google.dto.GoogleUserResponse;
-import com.woowacourse.momo.globalException.exception.MomoException;
+import com.woowacourse.momo.global.exception.exception.MomoException;
 
 @Getter
 @RequiredArgsConstructor
