+++ backend/src/main/java/com/woowacourse/momo/global/config/AspectConfiguration.java
@@ -1,4 +1,4 @@
-package com.woowacourse.momo.config;
+package com.woowacourse.momo.global.config;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
@@ -6,9 +6,9 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
 
-import com.woowacourse.momo.logging.ExceptionLogging;
-import com.woowacourse.momo.logging.LogFileManager;
-import com.woowacourse.momo.logging.Logging;
+import com.woowacourse.momo.global.logging.ExceptionLogging;
+import com.woowacourse.momo.global.logging.LogFileManager;
+import com.woowacourse.momo.global.logging.Logging;
 
 @Configuration
 @EnableAspectJAutoProxy
