+++ backend/src/main/resources/application-datasource.yml
@@ -1,4 +1,4 @@
 spring:
   jpa:
     hibernate:
-      naming.physical-strategy: com.woowacourse.momo.config.CustomNamingStrategy
+      naming.physical-strategy: com.woowacourse.momo.global.config.CustomNamingStrategy
